Add files via upload
R script for the analyses of "Community assembly of serpentine plants is driven by interspecific differences in functional traits." by G. Delhaye, P. G. Dimitrakopoulos and G. C. Adamidis. Script_analyses.R contains the main script and Script_FD.R only contains functions that are used in the main script. Figures.R uses the outputs of the script to produce the figures. Data are available upon request to the autors.